+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2018-04-26
+
 ### Added
 
 * Add the component dynamic schema props to hide the `showButtonOnHover` property if `hideBuyButton` is activated.